+++ arez/src/main/java/react4j/arez/ReactArezComponent.java
@@ -92,7 +92,7 @@ public abstract class ReactArezComponent
       _renderDepsChanged = true;
       if ( !_unmounted )
       {
-        scheduleRender( true );
+        scheduleRender();
       }
     }
   }